Add files via upload
Implemented node and matrix classes. Main methods included addElement() to add elements in any order, printMatrix() to represent the sparse matrix in visual form, and readFile() to read in a sparse matrix file, which uses addElement(). Main file is mostly empty and contains a few lines to test matrix class.